+++ app/Views/pages/home.php
@@ -1,6 +1,6 @@
-<?php /** @var array $prochainsMatchs */ ?>
+<?php /** @var array $prochainsMatchs @var array $matchsRecents */ ?>
 
-<?php echo partial('partials/upcoming_matches', ['prochainsMatchs' => $prochainsMatchs]); ?>
+<?php echo partial('partials/upcoming_matches', ['prochainsMatchs' => $prochainsMatchs, 'matchsRecents' => $matchsRecents]); ?>
 
 <div id="questions">
     <ul>
@@ -32,7 +32,7 @@
 <div id="numbers">
     <p>Highlander France, c'est :</p>
     <ul>
-        <li><span><b>+160</b></span> membres actifs</li>
+        <li><b><span>+</span><span id="memberCount">160</span></b> membres actifs</li>
         <li><b><span id="matchCount"><img class="stat-load" src="/_img/loading.gif" alt="Chargement..."></span></b> matchs organisés</li>
         <li><b><span>+</span><span id="hoursPlayed"><img class="stat-load" src="/_img/loading.gif" alt="Chargement..."></span></b> heures de matchs jouées au total</li>
     </ul>